chore(): eliminate unused files after refactoring
