+++ src/app/api/routes/read_pgn/__init__.py
@@ -1,0 +1,4 @@
+from .endpoints import router as read_pgn_router
+from .schemas import ReadPGNResponse
+
+__all__ = ["ReadPGNResponse", "read_pgn_router"]
