+++ src/app/api/routes/read_pgn/schemas.py
@@ -3,12 +3,6 @@ from pydantic import Field
 from app.api.schema import BaseSchema
 
 
-class UploadResponse(BaseSchema):
-    message: str = Field(description="Status message")
-    total_features: int = Field(ge=0, description="Number of features")
-    total_rows: int = Field(ge=0, description="Number of rows")
-
-
 class ReadPGNResponse(BaseSchema):
     message: str = Field(description="Status message")
     total_features: int = Field(ge=0, description="Number of features")
