--- src/app/api/routes/upload/__init__.py
@@ -1,4 +1,0 @@
-from .endpoints import router as upload_router
-from .schemas import ReadPGNResponse, UploadResponse
-
-__all__ = ["ReadPGNResponse", "UploadResponse", "upload_router"]
