+++ src/app/services/__init__.py
@@ -1,19 +1,15 @@
 from .exceptions import DataValidationError
 from .helper import load_model, save_model
 from .prediction import NoTrainedModelError, PredictionService
-from .preprocessing import PreprocessingService
 from .read_file import ReadFileService
 from .training import TrainingService
-from .upload import UploadService
 
 __all__ = [
     "DataValidationError",
     "NoTrainedModelError",
     "PredictionService",
-    "PreprocessingService",
     "ReadFileService",
     "TrainingService",
-    "UploadService",
     "load_model",
     "save_model",
 ]
