--- src/app/services/preprocessing/__init__.py
@@ -1,3 +1,0 @@
-from .service import PreprocessingService
-
-__all__ = ["PreprocessingService"]
