--- src/app/services/upload/__init__.py
@@ -1,8 +1,0 @@
-from app.domain.ml_model import MLModel
-
-from .service import UploadService
-
-__all__ = [
-    "MLModel",
-    "UploadService",
-]
